+++ src/main/java/vn/hactanco/socialnetwork/dto/UserSuggestionResponseDTO.java
@@ -1,0 +1,19 @@
+package vn.hactanco.socialnetwork.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+public class UserSuggestionResponseDTO {
+
+	private Long userId;
+	private String avatar;
+	private String name;
+}