+++ src/main/java/vn/hactanco/socialnetwork/enums/FriendshipStatus.java
@@ -1,0 +1,5 @@
+package vn.hactanco.socialnetwork.enums;
+
+public enum FriendshipStatus {
+	PENDING, ACCEPTED, REJECTED
+}