+++ src/main/java/vn/hactanco/socialnetwork/model/Friendship.java
@@ -1,0 +1,56 @@
+package vn.hactanco.socialnetwork.model;
+
+import java.time.Instant;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.PrePersist;
+import jakarta.persistence.Table;
+import jakarta.persistence.UniqueConstraint;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+import vn.hactanco.socialnetwork.enums.FriendshipStatus;
+
+@Entity
+@Table(name = "friendships", uniqueConstraints = { @UniqueConstraint(columnNames = { "follower_id", "following_id" }) })
+@Getter
+@Setter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Friendship {
+
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Long id;
+
+	// người gửi lời mời
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "follower_id", nullable = false)
+	private User follower;
+
+	// người nhận lời mời
+	@ManyToOne(fetch = FetchType.LAZY)
+	@JoinColumn(name = "following_id", nullable = false)
+	private User following;
+
+	@Enumerated(EnumType.STRING)
+	@Column(nullable = false)
+	private FriendshipStatus status;
+
+	private Instant createdAt;
+
+	@PrePersist
+	public void prePersist() {
+		this.createdAt = Instant.now();
+	}
+}